Fix Firebase null office data bug

+++ app/src/main/java/pl/droidsonroids/droidsmap/feature/office/business_logic/OfficeEntity.kt
@@ -2,4 +2,9 @@ package pl.droidsonroids.droidsmap.feature.office.business_logic
 
 import pl.droidsonroids.droidsmap.model.Entity
 
-data class OfficeEntity(val centerLatitude: Float = 0.0f, val centerLongitude: Float = 0.0f, val leftTopCornerLatitude: Float = 0.0f, val leftTopCornerLongitude: Float = 0.0f) : Entity()
+data class OfficeEntity(
+        val centerLatitude: Double = 0.0,
+        val centerLongitude: Double = 0.0,
+        val leftTopCornerLatitude: Double = 0.0,
+        val leftTopCornerLongitude: Double = 0.0
+) : Entity()
