+++ app/src/main/java/pl/droidsonroids/droidsmap/feature/office/mvp/OfficeUiModel.kt
@@ -2,7 +2,11 @@ package pl.droidsonroids.droidsmap.feature.office.mvp
 
 import pl.droidsonroids.droidsmap.feature.office.business_logic.OfficeEntity
 
-class OfficeUiModel private constructor(val centerLatitude: Float, val centerLongitude: Float, val leftTopCornerLatitude: Float, val leftTopCornerLongitude: Float) {
+class OfficeUiModel private constructor(
+        val centerLatitude: Double,
+        val centerLongitude: Double,
+        val leftTopCornerLatitude: Double,
+        val leftTopCornerLongitude: Double) {
     companion object {
         fun from(entity: OfficeEntity) = OfficeUiModel(
                 entity.centerLatitude,
